process listing working

+++ Userland/SHELL/include/stdio.h
@@ -4,7 +4,7 @@
 int putchar(char c);
 void printFF(const char * format, char ** s, int * n);
 void putNumber(int n);
-void * malloc(int index);
+void * malloc(long int index);
 int scanFF(const char * format, char ** s, int * n);
 int strcmpN(char * str1, char * str2,int number);
 int strcmp(char * str1, char * str2);
